Use SCons for builds instead of make
For now SCons replaces the Unix Makefile. It has all the same or
superior functionality.

For now the output files are always generated in "build", regardless
of the kind of build (console, debug, etc). Now however that SCons
is smart enough to rebuild everything when any build option changes,
so this is completely safe.

How to build:
  scons -h # to print the possible configuration options
  scons    # to build
  scons install # to install
  scons -c # to clean

The "install" target is not really useful yet.

TODO:
  - build different versions in different directories
  - build more than one version at a time
  - configure-like functionality for libraries
  - proper install target
  - mingw32 cross compilation support, Win32 mingw and cygwin support
  - MSVC support

I have an idea how to do all these things and they are not hard.