heapg.c,assertsc.c: check for HAVE_CRTDBG_H and HAVE_LIBDBGHELP
crtdbg.h and dbghelp.dll are not available under MingW and even not
always under MSVC. Changed the source to use them only if the
corresponding macro is defined.

Backported the USE_DBGHELPLIB from the master branch, but changed it
to HAVE_LIBDBGHELP to be consistent with established conventions and
used it in a couple of more places.