global.h: mingw definitions of __assume, __try and __except
The old source predicated the usage of __assume on WIN32. GCC doesn't
support it, so I predicated __assume on _MSC_VER >= 1000. Note that
while it may appear that GCC's __builtin_expect() can be used instead,
it actually can't. __builtin_expect() only works for branch prediction
in an "if".

Introduced a new feature macros HAVE_WIN32EH. If it is not defined,
__try and __except default to stubs. Exceptions won't be handled
properly, but at least it compiles.