Use bison+flex to generate files for fpcalc
Until now the pre-generated files (which I made a very long time ago
under Cygwin) were simply included in the source archive and there was
no need for bison+flex when compiling.

Now I have done the right thing -> the .l and .y files are processed by
bison+flex as they should. However, the original generated files are
still kept in the source (with .gen appended to their names), and if
bison or flex is missing the pregenerated file is used.

Currently there is no way to force using of the pregenerated file, but
it can be added should the need arise.